+++ packages/controller-utils/src/create-service-policy.ts
@@ -117,9 +117,7 @@ export type ServicePolicy = IPolicy & {
    * never succeeds before the retry policy gives up and before the maximum
    * number of consecutive failures has been reached.
    */
-  onDegraded: CockatielEvent<
-    FailureReason<unknown> | { duration: number } | void
-  >;
+  onDegraded: CockatielEvent<FailureReason<unknown> | { duration: number }>;
   /**
    * A function which is called when the service succeeds for the first time,
    * or when the service fails enough times to cause the circuit to break and
@@ -324,7 +322,7 @@ export function createServicePolicy(
   const onBreak = circuitBreakerPolicy.onBreak.bind(circuitBreakerPolicy);
 
   const onDegradedEventEmitter = new CockatielEventEmitter<
-    FailureReason<unknown> | { duration: number } | void
+    FailureReason<unknown> | { duration: number }
   >();
   const onDegraded = onDegradedEventEmitter.addListener;